always declare environ
POSIX says the user has to declare environ:

https://pubs.opengroup.org/onlinepubs/9699919799/
```
In addition, the following variable, which must be declared by the user
if it is to be used directly:

extern char **environ;
```

It would be really silly for systems to break this, I don't know of any
that do. C allows multiple declarations, so it's fine to do this on
systems that do declare environ for the user like glibc.